feat: Add permission to delete last invoice (#38)
* feat: Add createdByUserId to Invoice schema

* feat: Add permission for delete last invoice
- A user with the member role can only delete their last invoice.
- Warn members in modal window that they can only delete their own invoice.
- Display an error under the button after an unsuccessful deletion.

+++ app/api/invoices/last/route.ts
@@ -44,6 +44,19 @@ export async function DELETE() {
       );
     }
 
+    if (
+      membership.role === "member" &&
+      lastInvoice.createdByUserId !== session.user.id
+    ) {
+      return NextResponse.json(
+        {
+          error:
+            "Die letzte Rechnung wurde von einem anderen Benutzer erstellt und kann daher nicht gelöscht werden.",
+        },
+        { status: 403 },
+      );
+    }
+
     await prisma.item.deleteMany({
       where: { invoiceId: lastInvoice.id },
     });
